utility functions and PEP-8
added .ci/util.sh as this is the basis for the rest. It has all of the helper functions in it need to make check_*.sh and build_*.sh simpler to read.

add .ci/check_01.sh, which is that code meets PEP-8. Wronte some dummy stuff not in the repo to show that it works.