no issue: Simple readme to check the githooks with (#13)
* Updated the CI

Managed to connect Jenkins to github-fn. Update the dawgie-bot keys (deploy keys) which seems to be reasonably strong. All of the Jenkin's requests are run on mentor3.

Found some typos the CI checks. Cleaned those up as well.

Lastly, to allow Jenkins to send messages back to the webhook event, had to update my token. Took a while of digging to get it done, but it is now set. Messages from the Jenkins now appear correctly.

Completed the test with @groudier doing a review and approving the changes. The squash and merge turned green at that point regardless of the check 1 and 2 failures. Since check 5 passed, it is considered good enough today.