add dependency information
The first step in deploying excalibur as a docker image is to define which image it is dependent upon. In this case, it is on some version of dawgie and it will be tagged as such locally. The path element in artifactory allows for a more complete name in the scope of all docker images.