working docker builds
Dockerfile.py loads all of the dependencies for the excalibur system. Used find and grep to locate all of the imports and then eyeballed it to locate all of the external entity (not delivered with Python 3.6). Hence I may have missed a few but they should show up later in the pylint checks.

Dockerfile.cit is the CI tools for checking the quality of the excalibur code. It is dependent on the output of Dockerfile.py.