Issue 43: dockerize excalibur for deployment (#44)
* add dependency information

The first step in deploying excalibur as a docker image is to define which image it is dependent upon. In this case, it is on some version of dawgie and it will be tagged as such locally. The path element in artifactory allows for a more complete name in the scope of all docker images.

* working docker builds

Dockerfile.py loads all of the dependencies for the excalibur system. Used find and grep to locate all of the imports and then eyeballed it to locate all of the external entity (not delivered with Python 3.6). Hence I may have missed a few but they should show up later in the pylint checks.

Dockerfile.cit is the CI tools for checking the quality of the excalibur code. It is dependent on the output of Dockerfile.py.

* updated docker environment

.ci/Dockerfile.py updated to create directories that are world read-writable to anyone using the container. My guess is that these directories will need real content in the future but this is the minimum.

.ci/util.sh clean up a function that was finally excersized and needed a bit of touch up.

* code changes

Two actual code changes were required to go from pymc to pymc3. It seems deeper changes are required than just the naming so adding future calls should not be necessary at this commit.

* fix up pep-8 problems

.ci/Dockerfile.py is not a python function so should be excluded from coding standard checks.

Turns out that Python 3.6 is more rigorous about escaped characters. \mu should really be \\mu and will show up in the string as \mu in the end. Other minor white space issues.

* build the server to run on mentor0

There is not much beyond to this docker image beyond setting up all of the start parameters to shorten that part of the command line.

* New models in Atmos, Cerberus fully operationnal, Excalibur rated 9.02/10 by pylint, not too too bad for a code done by a scientist, I have to tell Al...

* build the worker containers

Added a setup.py so that excalibur can be installed in the container.
Added .ci/deploy_02.sh to build the worker image.

* put the git rev ID in setup.py as a version

* have a curl that should work in the deployment script