Minor change to try to resync workers (#62)
* add dependency information

The first step in deploying excalibur as a docker image is to define which image it is dependent upon. In this case, it is on some version of dawgie and it will be tagged as such locally. The path element in artifactory allows for a more complete name in the scope of all docker images.

* added some documentation since I have to look it up every time

* rename dependency file to make it match other tools

* initial docker files and build script

* wrong FROM PY layer

* add some meat to the dockerfiles

* working docker builds

Dockerfile.py loads all of the dependencies for the excalibur system. Used find and grep to locate all of the imports and then eyeballed it to locate all of the external entity (not delivered with Python 3.6). Hence I may have missed a few but they should show up later in the pylint checks.

Dockerfile.cit is the CI tools for checking the quality of the excalibur code. It is dependent on the output of Dockerfile.py.

* updated docker environment

.ci/Dockerfile.py updated to create directories that are world read-writable to anyone using the container. My guess is that these directories will need real content in the future but this is the minimum.

.ci/util.sh clean up a function that was finally excersized and needed a bit of touch up.

* dockerized the CI checks

* code changes

Two actual code changes were required to go from pymc to pymc3. It seems deeper changes are required than just the naming so adding future calls should not be necessary at this commit.

* fix up pep-8 problems

.ci/Dockerfile.py is not a python function so should be excluded from coding standard checks.

Turns out that Python 3.6 is more rigorous about escaped characters. \mu should really be \\mu and will show up in the string as \mu in the end. Other minor white space issues.

* turn on PEP-8 as a requirement

* make pymc/pymc3 compatible with either mentor or docker

* build the server to run on mentor0

There is not much beyond to this docker image beyond setting up all of the start parameters to shorten that part of the command line.

* build the worker containers

Added a setup.py so that excalibur can be installed in the container.
Added .ci/deploy_02.sh to build the worker image.

* put the git rev ID in setup.py as a version

* try to prevent install as egg

* fix up the worker and it actually seems to be working

* fix pep8 typo

* Fixed Formatting for ignored frames in data.cal that triggered error in transit.norm RID 1

* wild clean up of pylint errors

* wild clean up of cerberus

* wild cleanup of target and friends

* wild last bit

* Fixed data.calibration 1214 exception

* Aesthetic change in ransit core

* Fixed Broken Code in cerberus

* Passes CI2

* Added __main__.py in all tasks

* Passes CI1

* Fixed typo in transit.spectrum, HAT-P-17 parameters, added data.calibration truncated spectrum handling, Fixed HAT-P-41 spurious source exception

* Open the gate to STIS data

* Got rid of security check allowing only ima files to be ingested

* Passes CI2

* Fixed filter name for STIS data + target and system now on logger

* data task under logging + Fixed deterministic model in MCMC + create active filter for STIS

* White light with longer chains

* Created wavelength cal for STIS + Fixed Cerberus branching + First stab at calibrated stare mode data

* Fixed WFC3 ingest, STIS ingest, solves all issues related to compacted .fits files, MCMC OFF

* Trying to get workers back