Issue78 (#86)
* removed spurious states.py in eclipse

* complete inheritance

excalibur.transit.algirithms updating classes whitelight() and spectrum() to change two of the internal attributes to be protected (_) rather than private (__). This allows the child classes to redefine to the needs of the child and allow the child factories to instatiate them correctly.

excalibur.eclipse.algorithms updating classes whitelight() and spectrum() to use the classes defined here rather than the parent ones. As they were defined previously, the child classes would not have been able to override the parents value.

* Fixed elemental phase change handling in cerberus model, speed up whitelight, speed up transit, logging cosmetics