real changes to make science code more location independent
Added data_dir for the base data directory to find extraneous data.

Modified all paths to use os.path.join (excalibur.context['some location'], filename) instead of hard coding to /proj/sdp/data as it is no longer relavent in the container and will become less so in the cloud.