Issue100 (#103)
* Update deployment scripts

These changes are required to remove the tags from docker images in use then install the new tags. A delay was added to allow docker processes to stop and clean up before a sweep is made to delete the images that are now <none> for a tag.

There is still some spam in the screens running crew.sh. Removed the spam generator as well.