Issue208: aspect of regression (#216)
* Update to latest dawgie

Need fixes for the scaling to get runID back into the logs via version number changes.

Need changes to scheduler to support new request of an aspect of regressions.

* update target framework

Added the regression target.variations_of to read data from the target.autofill.parameters state vector. The skeletal structure is required now so that notebook page can retrieve the data for developing the run() action.

While expanding the structure for regression, added the aspect as well. The aspect works on the information generated by the regression.

* fill in the structure

Added excalibur.target.monitor to contain the new code for regression and aspect. Kept it out of core on purpose so that readers expect a slightly different coding style as well as content. core.py contains science code while monitor code is less science and more process.

Add excalibur.target.states.MonitorSV(). The original skeleton reused an existing SV but it was not appropriate for the monitor.

Updated excalibur.target.algorithms.Regression to include feedback(). This complicates things a tad but shows the use of it. Previously known infomration from the regression is passed into excalibur.target.monitor.regress() so that the loop over the timeline is as short as it can be. Normally a function like this would take more time on every cycle than it does when truncating at know run ids.

* fill in the alert

Worked through the aspect and filled in how it extracts the latest value from the regression and compares it to what it knows from last time. If it detects a change it will send an email or write an exception to the log. The aspect will always persist showing all of the latest parameters of concern for every target and and every planet of that target. A list will be at the top summarizing what the last changes are which is the same list that goes out via email.

* tested

Ran all of it in a notebook page. It works with the exception of email from the notebook. However email does seem to work otherwise. Might be something with the Jupyter container that is causing the issue. Put an exception block to handle the email failure.