Issue228: view aspect view (#229)
* add a fake visitor

The output is just a bunch of info that will not render images, tables, or text as they will be done in the pipeline, but it does allow one to easily run view() on a state vector to test that it will run to completion without crashing. If rendering is desired, then need to change excalibur.Visitor to render in a way that is friendly to the notebook.

* debugged view() in notebook

The results of debugging went quickly with the new visitor. Found a pretty simple error and fixed it.