tool for updating markdown
Completed the .ci/update_runids.py. Refactored a touch of it so that the body of the work is done a series of functions rather than in the __main__. It scans and reads data from the log files to extract run ids and changesets. It also reads what is currently in the markdown. New run IDs and changesets create new entries and then replaces the content of the markdown. It should be ready for integration into the deployment system.