prevent archive from hanging (#244)
Seems that PGPASSFILE is not defined for the server. This causes the archiving state to prompt for a password which then proceeds to hang forever. Added the ENV to the docker file so that it will always be set. The file is still located in sdppipded home directory and run_ops.sh is updated to include it as a mount. Tested using an -it docker container and it works again.