correct env usage (#277)
If you need this set, then define it in your .bash_profile or .cshrc or which ever shell you use. This will allow you to override it. If you need to override it in a container, use a -v ${HOME}/.ldtk:/proj/data/ldtk after you have defined the -v ...:/proj/data. Let me know if you need more help.