fix esp base build (#341)
* fix esp base build

Seems that the pip line had two problems. First, it had two references to matplotlib. One versioned and one not. The second was an unversioned numpy. Not versioning is a problem because it can cause code to break without being tested first.

* reinstall ubuntu python3-matplotlib

This is needed so that the next modules can find tkinter.