JWST + Targets + Spitzer Phasecurve (#354)
* ci checks + phase curve algorithms

* fix bad conflict merge (most likely)

* incorrect fix for dynamic attributes

* versions

* ci check fix; dynamic attributes

* fix pylint errors

The persistent errors are from generated attributes. Two problems with this are the rename with a from-import-as that later confounds pylint if it has been added to generate-members in pylintrc and the generated members to start with.

First, added the full name to the pylintrc. Second, went through phasecurve.core and data.core removing the renames as the shortnames with the canonical names.

* new target list

Co-authored-by: Kyle A. Pearson <kyle.a.pearson@nasa.jpl.gov>
Co-authored-by: Albert F Niessner <Albert.F.Niessner@jpl.nasa.gov>
Co-authored-by: Al Niessner <Al.Niessner@jpl.nasa.gov>
Co-authored-by: Gael M Roudier <Gael.M.Roudier@jpl.nasa.gov>
Co-authored-by: Noah Huber-Feely <noah.l.huber-feely@jpl.nasa.gov>