issue 439 injection (#440)
* newest matplotlib is less repetitive

* inject taurex spectrum

Using GJ 1214, injected a spectrum and changed the target name to 'GJ 1214(taurex something something)'. This required a framework for taurex that will allow for its expansion to more targets and different places in the pipeline. It mirrors that developed by @groudier in the rest of excalibur for easier consumption by those used to that coding style.

Used the new ds.retarget() ability to return not just these items as the new target name but historical items too. In this specific case system.finalize.

Lastly, connected cerberus to the state vectors hopefully in a meaningful way.