Issue 450: better code template (#451)
* remove the evil parts

The original idea was for just one target to go through the exercise but it required ugly and evil code to support it. Decided to just let all the targets make their way through the system and not worry with it. Code is much cleaner and more appropriate for an example to template.

* update to latest

Co-authored-by: Al Niessner <Al.Niessner@jpl.nasa.gov>