better tools to help (#643)
Problems with keeping the XSD synchronized with the data binding is that git can change the date/time for creation and modification during the cloning and merging process but the files may be unchanged. Instead, using the MD5 checksum to determine when they become out of sync and a tool to do the binding generation and create the checksums. Should be more robust even if an adveserial player can fake it.