Script upgrades (#703)
* script updates

The old scripts and security initializations incorrectly used PGP instead of TLS. The TLS is ignored when myself is not defined. Expanded the scripts to define and use the myself. Went through all of the __main__ and updtaed them as well to use myself correctly.

* Update pp_reset.sh

Bring it into line with start and stop after script upgrades. This is independent so can be merged independently.