docker cleanup (#6)
Needed to add the ESP git changeset to the worker so that it can appropriately connect to the pipeline.

Set a default value to base to make docker compose happy. Despite it being defined everywhere in the compose file, build still complains that an ARG does not have a default value. They do now.